+++ cli/cli/commands/files/storeservice/storeservice.go
@@ -89,7 +89,7 @@ func run(
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the enclave context for enclave '%v'", enclaveId)
 	}
-	filesArtifactId, err := enclaveCtx.StoreFilesFromUserService(ctx, serviceId, filepath)
+	filesArtifactId, err := enclaveCtx.StoreFilesFromService(ctx, serviceId, filepath)
 	if err != nil {
 		return stacktrace.Propagate(
 			err,