+++ cli/internal_testsuites/golang/testsuite/exec_command_test/exec_command_test.go
@@ -17,6 +17,8 @@ const (
 	execCmdTestImage      = "alpine:3.12.4"
 	inputForLogOutputTest = "hello"
 	expectedLogOutput     = "hello\n"
+	inputForAdvancedLogOutputTest = "hello && hello"
+	expectedAdvancedLogOutput = "hello && hello\n"
 	testServiceId         = "test"
 
 	successExitCode int32 = 0
@@ -30,6 +32,13 @@ var execCommandThatShouldWork = []string{
 var execCommandThatShouldHaveLogOutput = []string{
 	"echo",
 	inputForLogOutputTest,
+}
+
+// This command tests to ensure that the commands the user is running get passed exactly as-is to the Docker
+// container. If Kurtosis code is magically wrapping the code with "sh -c", this will fail.
+var execCommandThatWillFailIfShWrapped = []string{
+	"echo",
+	inputForAdvancedLogOutputTest,
 }
 
 var execCommandThatShouldFail = []string{
@@ -51,23 +60,31 @@ func TestExecCommand(t *testing.T) {
 	require.NoError(t, err, "An error occurred starting service '%v'", testServiceId)
 
 	// ------------------------------------- TEST RUN ----------------------------------------------
-	logrus.Infof("Running exec command '%v' that should return a successful exit code...", execCommandThatShouldWork)
+	logrus.Infof("Running exec command '%+v' that should return a successful exit code...", execCommandThatShouldWork)
 	shouldWorkExitCode, _, err := runExecCmd(testServiceContext, execCommandThatShouldWork)
-	require.NoError(t, err, "An error occurred running exec command '%v'", execCommandThatShouldWork)
-	require.Equal(t, successExitCode, shouldWorkExitCode, "Exec command '%v' should work, but got unsuccessful exit code %v", execCommandThatShouldWork, shouldWorkExitCode)
+	require.NoError(t, err, "An error occurred running exec command '%+v'", execCommandThatShouldWork)
+	require.Equal(t, successExitCode, shouldWorkExitCode, "Exec command '%+v' should work, but got unsuccessful exit code %v", execCommandThatShouldWork, shouldWorkExitCode)
 	logrus.Info("Exec command returned successful exit code as expected")
 
-	logrus.Infof("Running exec command '%v' that should return an error exit code...", execCommandThatShouldFail)
+	logrus.Infof("Running exec command '%+v' that should return an error exit code...", execCommandThatShouldFail)
 	shouldFailExitCode, _, err := runExecCmd(testServiceContext, execCommandThatShouldFail)
-	require.NoError(t, err, "An error occurred running exec command '%v'", execCommandThatShouldFail)
-	require.NotEqual(t, successExitCode, shouldFailExitCode, "Exec command '%v' should fail, but got successful exit code %v", execCommandThatShouldFail, successExitCode)
+	require.NoError(t, err, "An error occurred running exec command '%+v'", execCommandThatShouldFail)
+	require.NotEqual(t, successExitCode, shouldFailExitCode, "Exec command '%+v' should fail, but got successful exit code %v", execCommandThatShouldFail, successExitCode)
+	logrus.Infof("Exec command returning an error exited with error")
 
-	logrus.Infof("Running exec command '%v' that should return log output...", execCommandThatShouldHaveLogOutput)
+	logrus.Infof("Running exec command '%+v' that should return log output...", execCommandThatShouldHaveLogOutput)
 	shouldHaveLogOutputExitCode, logOutput, err := runExecCmd(testServiceContext, execCommandThatShouldHaveLogOutput)
-	require.NoError(t, err, "An error occurred running exec command '%v'", execCommandThatShouldHaveLogOutput)
-	require.Equal(t, successExitCode, shouldHaveLogOutputExitCode, "Exec command '%v' should work, but got unsuccessful exit code %v", execCommandThatShouldHaveLogOutput, shouldHaveLogOutputExitCode)
-	require.Equal(t, expectedLogOutput, logOutput, "Exec command '%v' should return %v, but got %v.", execCommandThatShouldHaveLogOutput, inputForLogOutputTest, logOutput)
-	logrus.Info("Exec command returned error exit code as expected")
+	require.NoError(t, err, "An error occurred running exec command '%+v'", execCommandThatShouldHaveLogOutput)
+	require.Equal(t, successExitCode, shouldHaveLogOutputExitCode, "Exec command '%+v' should work, but got unsuccessful exit code %v", execCommandThatShouldHaveLogOutput, shouldHaveLogOutputExitCode)
+	require.Equal(t, expectedLogOutput, logOutput, "Exec command '%+v' should return '%v', but got '%v'.", execCommandThatShouldHaveLogOutput, expectedLogOutput, logOutput)
+	logrus.Info("Exec command returning log output passed as expected")
+
+	logrus.Infof("Running exec command '%+v' that will fail if Kurtosis is accidentally sh-wrapping the command...", execCommandThatWillFailIfShWrapped)
+	shouldNotGetShWrappedExitCode, shouldNotGetShWrappedLogOutput, err := runExecCmd(testServiceContext, execCommandThatWillFailIfShWrapped)
+	require.NoError(t, err, "An error occurred running exec command '%+v'", execCommandThatWillFailIfShWrapped)
+	require.Equal(t, successExitCode, shouldNotGetShWrappedExitCode, "Exec command '%v' should work, but got unsuccessful exit code %v", execCommandThatWillFailIfShWrapped, shouldNotGetShWrappedExitCode)
+	require.Equal(t, expectedAdvancedLogOutput, shouldNotGetShWrappedLogOutput, "Exec command '%+v' should return '%v', but got '%v'.", execCommandThatWillFailIfShWrapped, expectedAdvancedLogOutput, shouldNotGetShWrappedLogOutput)
+	logrus.Info("Exec command that will fail if Kurtosis is accidentally sh-wrapping did not fail")
 }
 
 // ====================================================================================================