+++ cli/docs/changelog.md
@@ -1,4 +1,11 @@
 # TBD
+### Features
+* Added `files storeservice` to store files from a service in the enclave
+
+### Changes
+* Upgraded `kurtosis core lib` to 1.43.6 which adds `EnclaveContext.StoreFilesFromService` 
+* Upgraded `kurtosis engine lib` to 1.15.6
+* Upgraded `container engine lib` to 0.16.0
 
 # 0.14.1
 ### Features